Refactorized all classes Deleted Qurami Changed "Doctore Finder" Margin on X on Login and Signup added
