+++ Doctor-Finder-App/app/src/main/java/com/doctorfinderapp/doctorfinder/activity/Informativa.java
@@ -1,4 +1,4 @@
-package com.doctorfinderapp.doctorfinder;
+package com.doctorfinderapp.doctorfinder.activity;
 
 import android.graphics.Color;
 import android.os.Bundle;
@@ -12,6 +12,8 @@ import android.util.Log;
 import android.view.MenuItem;
 import android.view.View;
 import android.widget.Toast;
+
+import com.doctorfinderapp.doctorfinder.R;
 
 
 /**
