+++ Doctor-Finder-App/app/src/main/java/com/doctorfinderapp/doctorfinder/activity/WebViewActivity.java
@@ -1,4 +1,4 @@
-package com.doctorfinderapp.doctorfinder;
+package com.doctorfinderapp.doctorfinder.activity;
 
 import android.content.Intent;
 import android.net.Uri;
@@ -11,6 +11,8 @@ import android.webkit.WebSettings;
 import android.webkit.WebView;
 import android.webkit.WebViewClient;
 import android.widget.Toast;
+
+import com.doctorfinderapp.doctorfinder.R;
 
 
 public class WebViewActivity extends AppCompatActivity {
