+++ Doctor-Finder-App/app/src/main/java/com/doctorfinderapp/doctorfinder/activity/access/FirstActivity.java
@@ -1,10 +1,8 @@
-package com.doctorfinderapp.doctorfinder.access;
+package com.doctorfinderapp.doctorfinder.activity.access;
 
 import android.app.Activity;
 import android.app.AlertDialog;
 import android.content.Intent;
-import android.content.res.Resources;
-import android.graphics.Bitmap;
 import android.os.Bundle;
 import android.view.View;
 import android.view.WindowManager;
@@ -12,16 +10,8 @@ import android.widget.Button;
 import android.widget.ImageButton;
 import android.widget.TextView;
 
-import com.doctorfinderapp.doctorfinder.Intro.DefaultIntro;
-import com.doctorfinderapp.doctorfinder.MainActivity;
+import com.doctorfinderapp.doctorfinder.activity.MainActivity;
 import com.doctorfinderapp.doctorfinder.R;
-import com.facebook.internal.DialogPresenter;
-import com.parse.ParseUser;
-
-import cn.pedant.SweetAlert.SweetAlertDialog;
-import com.doctorfinderapp.doctorfinder.functions.AddDoctors;
-
-import java.security.Policy;
 
 public class FirstActivity extends Activity {
