+++ Doctor-Finder-App/app/src/main/java/com/doctorfinderapp/doctorfinder/activity/access/SplashActivity.java
@@ -1,4 +1,4 @@
-package com.doctorfinderapp.doctorfinder.access;
+package com.doctorfinderapp.doctorfinder.activity.access;
 
 
 import android.content.Intent;
@@ -6,8 +6,8 @@ import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
 import android.view.View;
 
-import com.doctorfinderapp.doctorfinder.Intro.DefaultIntro;
-import com.doctorfinderapp.doctorfinder.MainActivity;
+import com.doctorfinderapp.doctorfinder.activity.intro.DefaultIntro;
+import com.doctorfinderapp.doctorfinder.activity.MainActivity;
 import com.doctorfinderapp.doctorfinder.R;
 import com.doctorfinderapp.doctorfinder.functions.FacebookProfile;
 import com.parse.ParseUser;
