+++ Doctor-Finder-App/app/src/main/java/com/doctorfinderapp/doctorfinder/objects/Person.java
@@ -1,4 +1,4 @@
-package com.doctorfinderapp.doctorfinder.Class;
+package com.doctorfinderapp.doctorfinder.objects;
 
 
 public class Person {
