+++ Doctor-Finder-App/app/src/main/res/layout/activity_doctor.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
-    tools:context="com.doctorfinderapp.doctorfinder.DoctorActivity">
+    tools:context=".activity.DoctorActivity">
 
 
         <android.support.design.widget.AppBarLayout
