+++ Doctor-Finder-App/app/src/main/res/layout/activity_main.xml
@@ -14,7 +14,7 @@
         android:layout_width="match_parent"
         android:layout_height="match_parent"
         android:fitsSystemWindows="true"
-        tools:context=".MainActivity"
+        tools:context=".activity.MainActivity"
         android:background="@color/google_grey">
 
         <android.support.design.widget.AppBarLayout
