+++ Doctor-Finder-App/app/src/main/res/layout/activity_results.xml
@@ -7,7 +7,7 @@
     android:id="@+id/drawer_results"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ResultsActivity"
+    tools:context=".activity.ResultsActivity"
     android:fitsSystemWindows="true"
     android:background="@color/colorPrimaryDark">
 
@@ -17,7 +17,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ResultsActivity"
+    tools:context=".activity.ResultsActivity"
     android:clickable="true"
     android:focusable="true">
