+++ Doctor-Finder-App/app/src/main/res/layout/activity_signup.xml
@@ -15,7 +15,8 @@
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:id="@+id/close2"
-        android:background="@drawable/ic_close_white_24dp" />
+        android:background="@drawable/ic_close_white_24dp"
+        android:layout_marginTop="5dp" />
     </LinearLayout>
 
     <ImageView
