+++ Doctor-Finder-App/app/src/main/res/layout/activity_web_view.xml
@@ -9,7 +9,7 @@
     android:fitsSystemWindows="true"
 
 
-    tools:context="com.doctorfinderapp.doctorfinder.WebViewActivity">
+    tools:context=".activity.WebViewActivity">
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
