+++ Doctor-Finder-App/app/src/main/res/layout/intro.xml
@@ -45,7 +45,7 @@
         android:paddingRight="64dp"
         android:textSize="16sp"
         android:fontFamily="sans-serif-thin"
-        android:text="Doctore Finder ti permette di trovare il dottore giusto per te utilizzando i parametri che ritieni più giusti"/>
+        android:text="Doctor Finder ti permette di trovare il dottore perfett per te utilizzando i parametri che ritieni più giusti"/>
     <TextView
         android:layout_width="fill_parent"
         android:layout_height="64dp" />
