+++ Doctor-Finder-App/app/src/main/res/menu/menu_doctor.xml
@@ -1,7 +1,7 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="com.doctorfinderapp.doctorfinder.DoctorActivity">
+    tools:context=".activity.DoctorActivity">
     <item
         android:id="@+id/action_settings"
         android:orderInCategory="100"
